# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dinas PMD Katingan API Documentation",
        "description": "RESTful API untuk Website Dinas Pemberdayaan Masyarakat dan Desa Kabupaten Katingan.\n *                   API ini menyediakan akses terhadap data berita, agenda, galeri, dokumen, dan struktur organisasi.",
        "contact": {
            "name": "Dinas PMD Katingan",
            "email": "info@pmdkatingan.go.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server (Configurable via .env)"
        }
    ],
    "paths": {
...
        "/api/v1/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get list of published news",
                "description": "Returns paginated list of published news articles",
                "operationId": "getNewsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title and content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/news/{slug}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news detail by slug",
                "description": "Returns single news article with full content",
                "operationId": "getNewsDetail",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "News slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "News not found"
                    }
                }
            }
        },
...
        "/api/v1/organization/{id}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organization member detail",
                "description": "Returns detailed information of a specific organization member",
                "operationId": "getOrganizationMemberDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Member not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token obtained from login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}